+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 - Added --data-path to display the active application data directory.
 - Added --backup to create a local JERVIS data backup from the command line.
+- Added --list-backups to display available JERVIS data backups.
 ## [1.5.2] - 2026-09-01
 
 ### Changed